# docs/swagger.yaml
basePath: /
definitions:
  domain.ConfirmCode:
    properties:
      code:
        type: string
      email:
        type: string
    required:
    - code
    - email
    type: object
  domain.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      path:
        type: string
      timeStamp:
        type: string
      title:
        type: string
    type: object
...
  domain.ResetPassword:
    properties:
      confirm:
        minLength: 6
        type: string
      new:
        minLength: 6
        type: string
    required:
    - confirm
    - new
    type: object
  domain.UpdatePassword:
    properties:
      current:
        minLength: 6
        type: string
      new:
        minLength: 6
        type: string
    required:
    - current
    - new
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API para gerenciamento de autenticação, incluindo registro de usuários,
    login, confirmação de e-mail, e recuperação de senha.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Authentication Pulse Tech API
  version: "1.0"
paths:
...
  /v1/auth/password:
    post:
      consumes:
      - application/json
      description: Send an OTP code to redeem your password
      parameters:
      - description: Confirmation Code Payload
        in: body
        name: confirmCode
        required: true
        schema:
          $ref: '#/definitions/domain.ConfirmCode'
      produces:
      - application/json
      responses:
        "200":
          description: JWT Token
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Forgot user password
      tags:
      - authentication
  /v1/auth/password/confirm:
    post:
      consumes:
      - application/json
      description: Confirm the reset password code sent to the user's email
      parameters:
      - description: Confirmation Code
        in: body
        name: confirmCode
        required: true
        schema:
          $ref: '#/definitions/domain.ConfirmCode'
      produces:
      - application/json
      responses:
        "200":
          description: JWT Token
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Confirm reset password code
      tags:
      - authentication
...
  /v1/users/password/{id}:
    patch:
      consumes:
      - application/json
      description: Update password for authenticated users
      parameters:
      - description: Update Password Payload
        in: body
        name: updatePassword
        required: true
        schema:
          $ref: '#/definitions/domain.UpdatePassword'
      produces:
      - application/json
      responses:
        "200":
          description: JWT Token
          schema:
            type: string
        "403":
          description: Forbidden
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - bearerToken: []
      summary: Update password user
      tags:
      - users
  /v1/users/password/reset:
    patch:
      consumes:
      - application/json
      description: Reset the password for the authenticated user
      parameters:
      - description: Reset Password Data
        in: body
        name: resetPassword
        required: true
        schema:
          $ref: '#/definitions/domain.ResetPassword'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - bearerToken: []
      summary: Reset user password
      tags:
      - authentication
...
schemes:
- http
securityDefinitions:
  bearerToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"